MSVC++ portability--modified code to work around MS Visual C++ limitations and bugs.  Highlights:
  * Added (char) casts to static initializers to supress warnings.
  * Replaced 'int' with 'int32' when calling TString::operator[] to
    prevent weird overloading errors.
  * Made encoding error-logging functions take a size_t parameter
    instead of an int to represent string positions.
  * Changed code which assumed that the variable 'i' in 'for (int i'
    was scoped according to the C++ standard.  Everybody else fixed
    their C++ compilers about half a decade ago, but Microsoft never
    got arround to it.  Fun, fun.
  * Fixed miscellaneous portability problems in code.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@27144 b24aca02-f418-0410-9c49-835e4aa915c3

+++ Common/TEncoding.h
@@ -57,7 +57,7 @@ public:
 	// [in] inBadPos - the 0-based character position of the error
 	// [in] inErrMsg - a message explaining what is wrong
 	typedef void (*ErrorLoggingFunc) (const char *inBadString,
-									  int inBadPos,
+									  size_t inBadPos,
 									  const char *inErrMsg);
 
 private:
